thumbnail_analysis/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.0
google-auth==1.20.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
idna==2.10
pika==1.1.0
protobuf==3.12.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
